soc:arm: dts:arm:xilinx: SoC-specific code and device tree data for the Xilinx Zynq-7000
Add SoC-specific code, the basic device tree and Kconfig data as well as the corresponding
linker command file for the Xilinx Zynq-7000 family of SoCs. This SoC - either as a QEMU
simulation or on actual hardware such as the Avnet/Digilent ZedBoard - is suitable as an
initial target for the ARMv7 Cortex-A support.

Signed-off-by: Immo Birnbaum <Immo.Birnbaum@weidmueller.com>