arch:arm:core:aarch32: support for On-Chip Memory (OCM) areas in aarch32
Introduce support for On-Chip Memory (OCM) areas in aarch32 Cortex-A SoCs.
This memory area is suitable for, amongst other things, uncached device DMA.
The OCM memory can reside either at address zero or at a high memory address,
with configuration of the memory's location similar to the configuration of
the location of the exception vectors. In both cases, it must be assured
that the low/high OCM area declaration doesn't overlap with the exception
vectors memory page at either the low/high address.
Therefore, all possible OCM location options shall be described in the
SoC's device tree, the current location shall be indicated via a "chosen"
entry in the target board's device tree.

This commit contains the necessary changes to the linker-related files
as well as the static region declaration for the MMU initialization in
the Zynq-7000 SoC-specific init code.

Signed-off-by: Immo Birnbaum <Immo.Birnbaum@weidmueller.com>